Update code formatting and linting configs
The changes in this commit pertain to the configuration of the code linter and formatter, namely '.ruff.toml' and '.pre-commit-config.yaml'. Main adjustments include specifying line length and indentation width in '.ruff.toml', and moving formatting arguments to the 'ruff-format' hook in '.pre-commit-config.yaml'. The commit also reorganizes certain sections under the 'lint' category in '.ruff.toml'.